Move SheetContainer to a separate file
