+++ package/lib/src/draggable/draggable_sheet.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/widgets.dart';
 
-import '../foundation/framework.dart';
 import '../foundation/keyboard_dismissible.dart';
 import '../foundation/physics.dart';
+import '../foundation/sheet_container.dart';
 import '../foundation/sheet_controller.dart';
 import '../foundation/sheet_extent.dart';
 import '../foundation/theme.dart';
