+++ package/lib/src/scrollable/scrollable_sheet.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:meta/meta.dart';
 
-import '../foundation/framework.dart';
 import '../foundation/keyboard_dismissible.dart';
 import '../foundation/physics.dart';
+import '../foundation/sheet_container.dart';
 import '../foundation/sheet_controller.dart';
 import '../foundation/sheet_extent.dart';
 import '../foundation/theme.dart';
